Emua Project
Code Review, CategoryList Screen was updated.
