--- app/src/main/res/layout/product_list_product_format.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="horizontal" >
-
-    <ImageView
-        android:id="@+id/icon"
-        android:layout_width="150dp"
-        android:layout_height="150dp"
-		android:padding="5dp" />
-
-    <LinearLayout android:layout_width="wrap_content"
-        android:layout_height="150dp"
-        android:orientation="vertical"
-        android:layout_marginLeft="10dp"
-        android:layout_marginRight="10dp"
-        android:layout_marginTop="10dp">
-    
-        <TextView
-            android:id="@+id/item"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Medium Text"
-            android:textAppearance="?android:attr/textAppearanceMedium"
-            android:padding="5dp"
-            android:textSize="32sp"
-            android:textColor="#33CC33" />
-
-        <TextView
-            android:id="@+id/textView1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceMedium"
-            android:layout_gravity="center"
-            android:text="TextView"
-            android:textSize="25sp"
-            android:layout_marginTop="10dp"
-            android:padding="5dp"/>
-
-    </LinearLayout>
-
-</LinearLayout>
