+++ app/src/main/res/layout/single_item_in_category_list.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:padding="5dp"
+    android:orientation="horizontal">
+
+    <ImageView
+        android:id="@+id/grid_item_image"
+        android:layout_width="70dp"
+        android:layout_height="70dp"
+        android:layout_marginRight="20px">
+    </ImageView>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+
+        <TextView
+            android:id="@+id/category_name_label"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Category"
+            android:layout_marginTop="15px"
+            android:textSize="12dp"
+            android:textColor="#6ca74a">
+        </TextView>
+
+        <TextView
+            android:id="@+id/category_description_label"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Bu kategori altında 12 adet ürün vardır"
+            android:layout_marginTop="5px"
+            android:textSize="7dp" >
+        </TextView>
+    </LinearLayout>
+
+
+
+</LinearLayout>
